+++ src/filters.js
@@ -1,11 +1,7 @@
 export function isFiltered(action, localFilter) {
   if (typeof window === 'undefined' && !localFilter) return true;
-  if (
-    !localFilter && !window.devToolsOptions ||
-    !window.devToolsOptions.filter || window.devToolsOptions.filter === 'DO_NOT_FILTER'
-  ) return false;
-
-  const { whitelist, blacklist } = localFilter || window.devToolsOptions;
+  if (!localFilter) return false;
+  const { whitelist, blacklist } = localFilter;
   return (
     whitelist && !action.type.match(whitelist) ||
     blacklist && action.type.match(blacklist)
