+++ src/utils.js
@@ -57,3 +57,44 @@ function setValueAction(store, state) {
 setValueAction.__isRemotedevAction = true;
 export const setValue = mobx.action('@@remotedev', setValueAction);
 /* eslint-enable */
+
+export function getMethods(obj) {
+  if (typeof obj !== 'object') return undefined;
+  let functions;
+  let m;
+  if (obj.__proto__) m = obj.__proto__.__proto__;
+  if (!m) m = obj;
+
+  Object.getOwnPropertyNames(m).forEach(key => {
+    const prop = m[key];
+    if (typeof prop === 'function' && key !== 'constructor') {
+      if (!functions) functions = [];
+      functions.push({
+        name: key || prop.name || 'anonymous',
+        args: getParams(prop),
+      });
+    }
+  });
+  return functions;
+}
+
+/* eslint-disable no-new-func */
+export const interpretArg = (arg) => (new Function('return ' + arg))();
+
+export function evalArgs(inArgs, restArgs) {
+  const args = inArgs.map(interpretArg);
+  if (!restArgs) return args;
+  const rest = interpretArg(restArgs);
+  if (Array.isArray(rest)) return args.concat(...rest);
+  throw new Error('rest must be an array');
+}
+
+export function evalMethod(action, obj) {
+  if (typeof action === 'string') {
+    return (new Function('return ' + action)).call(obj);
+  }
+
+  const args = evalArgs(action.args, action.rest);
+  return (new Function('args', `return this.${action.name}(args)`)).apply(obj, args);
+}
+/* eslint-enable */
